test: replace mocha with uvu, reorganize test suite
- Swap mocha for uvu
- Split monolithic file into suites
