--- .mocharc.json
@@ -1,7 +1,0 @@
-{
-  "check-leaks": false,
-  "throw-deprecation": true,
-  "trace-deprecation": true,
-  "trace-warnings": true,
-  "use-strict": true
-}
